% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAccumulationBuffer.R
\name{calcAccumulationBuffer}
\alias{calcAccumulationBuffer}
\title{Buffering of flow accumulations using thresholds}
\usage{
calcAccumulationBuffer(in.grid, out.grid, out.shp = NULL,
  return.sf = FALSE, t, buf.dist, working.path = NULL,
  keep.intermediate = FALSE, noData = -99999,
  show.output.on.console = FALSE, quiet = TRUE)
}
\arguments{
\item{in.grid}{path of input grid (flow accumulation)}

\item{out.grid}{path of output grid}

\item{out.shp}{creation of an output shapefile. Default: NULL}

\item{return.sf}{return of simple feature. Default: FALSE}

\item{t}{vector containing thresholds for flow accumulation grid}

\item{buf.dist}{vector containing the buffer distances}

\item{working.path}{working path. Default: NULL}

\item{keep.intermediate}{keep intermediate data. Default: FALSE}

\item{noData}{value for no data input. Default: -99999}

\item{show.output.on.console}{show output on console. Default: FALSE}

\item{quiet}{no outputs in console. Default: TRUE}
}
\value{
if \emph{out.shp} is set and \emph{return.sf} is set to TRUE, then a simple feature object is returned
}
\description{
This function buffers a flow accumulation input in dependence on given thesholds and distances
}
\note{
length of \emph{t} and \emph{buf.dist} must be identical
}
\keyword{accumulation,}
\keyword{buffer}
\keyword{flow}
