% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkGeometry.R
\name{checkGeometry}
\alias{checkGeometry}
\title{Geometry check and repair of SpatialPolygonsDataFrame}
\usage{
checkGeometry(spdf, bug.correct.rgeos = TRUE, bug.simpl.tol = 1e-05,
  bug.buf.width = 0, quiet = TRUE)
}
\arguments{
\item{spdf}{\linkS4class{SpatialPolygonsDataFrame} input}

\item{bug.correct.rgeos}{correct rgeos bug, see \href{https://gis.stackexchange.com/questions/163445/r-solution-for-topologyexception-input-geom-1-is-invalid-self-intersection-er}{TopologyException} \emph{(last call: 13-04-2017)}. Default: TRUE}

\item{bug.simpl.tol}{tolerance in \code{\link[rgeos]{gSimplify}}. Default: 0.00001}

\item{bug.buf.width}{buffer width in \code{\link[rgeos]{gBuffer}}. Default: 0}

\item{quiet}{no outputs in console. Default: TRUE}
}
\value{
repaired \linkS4class{SpatialPolygonsDataFrame}
}
\description{
This function checks the geometry of a \linkS4class{SpatialPolygonsDataFrame} input. Invalid geometries are repaired by the \code{\link[cleangeo]{clgeo_Clean}} function.
}
\examples{
checkGeometry()

}
\keyword{cleangeo}
\keyword{geometry,}
\keyword{invalid}
\keyword{rgeos,}
