% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classNeighborFunction.R
\name{classNeighborFunction}
\alias{classNeighborFunction}
\title{Calculation of relations from a class to its neighbor objects}
\usage{
classNeighborFunction(spdf, nb, class.var, var, calc.nb.flow = FALSE,
  col.flow = NULL, bb.intersection = "rgeos", class.as.neighbors = FALSE,
  bb = NULL, return.bb.flow = FALSE, scale.factor.flow = c(1, 1),
  centroid.flow = TRUE, set.centroid.flow = "inverse",
  k.centroid.flow = 2, k.flow = 2, scale.side.flow = "small",
  quiet = TRUE)
}
\arguments{
\item{spdf}{\linkS4class{SpatialPolygonsDataFrame}, or full path of a shapefile}

\item{nb}{object neighborhood based on \code{\link[spdep]{poly2nb}}. Otherwise, neighborhood is computed by spdep::poly2nb(..., Queens = TRUE)}

\item{class.var}{defination of class. Can be string or integer, or an expression in form of a list. For examples: class = as.integer(1), class = list("> 2.5", "expression"), class = list(c("> 2", "&", "< 7"), "expression")}

\item{var}{vector with number or name defining 1. field of evaluation (i.e. slope) and 2. field of weights (i.e. area). For example: c("Slope", "Area") or c(1,2)}

\item{calc.nb.flow}{include statistics for neighbors in flow direction. Default: FALSE}

\item{col.flow}{column specifying direction. Default: NULL}

\item{bb.intersection}{method for intersection with bounding box. Default: "rgeos" (alternative: "sp")}

\item{class.as.neighbors}{class object is also neighbor object. Default: FALSE}

\item{bb}{input of bounding box geometry. Default: NULL}

\item{return.bb.flow}{return geoemtry of counding box. Default: FALSE}

\item{quiet}{no outputs in console. Default: TRUE}

\item{only.neighbors}{object statistics are computed only for neighbors. Default: TRUE (FALSE means computation for the entiry data)}

\item{other...}{see \code{\link{getBoundingBox}}}
}
\value{
\linkS4class{data.frame} containing object statistics, and optional the created geometry of \code{\link{getBoundingBox}}
}
\description{
This function calculates statistics for every object of a specific class to its neighbor objects.
It is computationally identical to the \code{\link{relationalObjectFunction}}.
However, the function additionally enables the possibility to perform statistics for neighbors in a specific direction,
likewise the \code{\link{relationalClassFunction}}, and to select class objects by expressions.
}
\note{
\itemize{
  \item class statistics to neighbor objects acccording to ECOGNITION DEVELOPER (2014: 253-255, 375-381):
  \itemize{
      \item see \code{\link{relationalObjectFunction}}
      \item \emph{~_nb_FlAl} - statistics for all objects located in flow direction
      \item \emph{~_nb_FlMa} - statistics for direct neighbors located in flow direction
      }
  \item ECOGNITION DEVELOPER (2014) Reference Book. Trimble Documentation, München, Germany
  \item see \code{\link{getBoundingBox}}
  \item \linkS4class{SpatialPolygonsDataFrame} objects with NA values are ignored
  \item \linkS4class{SpatialPolygonsDataFrame} objects without neighbors are ignored
  \item \linkS4class{SpatialPolygonsDataFrame} input must have a valid \emph{Class} and \emph{ID} field.
     \emph{ID} field must be contained of unique numbers
}
}
\keyword{analysis}
\keyword{class}
\keyword{image}
\keyword{neighbor}
\keyword{object-oriented}
\keyword{objects}
\keyword{objects,}
\keyword{of}
\keyword{relation}
\keyword{spdep,}
\keyword{their}
\keyword{to}
