% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrastFilterSegmentation.R
\name{contrastFilterSegmentation}
\alias{contrastFilterSegmentation}
\title{Contrast-Filter-Segmentation}
\usage{
contrastFilterSegmentation(input.filter, input.segmentation = input.filter,
  offset = 0.06, makeBrush.size = 25, makeBrush.shape = "disc",
  NA.val.in = 0, clump.thresh = NULL, clump.directions = 4,
  env.rsaga = RSAGA::rsaga.env(), CFR.buf = 1, output.path = tempdir(),
  writeCFRaster = FALSE, writeRaster.NAflag = -99999, freeMemory = TRUE,
  show.output.on.console = FALSE, quiet = TRUE, morph.Closing = TRUE,
  closing.size = 3, closing.shape = "box",
  Grass.Segmentation.Threshold = 0.24, Grass.Segmentation.Minsize = 1,
  Grass.Segmentation.Memory = 1024, Segments.Poly = file.path(tempdir(),
  "outSegPoly.shp"), Segments.Grid = file.path(tempdir(), "outSegGrid.sgrd"),
  defaultGrass = c("C:/OSGeo4W64", "grass-7.2.2", "OSGeo4W64"),
  load.output = FALSE, fill.holes = FALSE, ...)
}
\arguments{
\item{input.filter}{\linkS4class{RasterLayer} on which the filtering is applied}

\item{input.segmentation}{\linkS4class{RasterLayer} for segmentation. Default: input.filter}

\item{offset}{Offset of values in filter matrix. Default: 0.06}

\item{makeBrush.size}{matrix size for smoothing input.filter. Default: 25}

\item{makeBrush.shape}{form of smoothin matrix. Default: "disc"}

\item{NA.val.in}{replace value for no data (NA) in \emph{input.filter}. Default: 0}

\item{clump.thresh}{area thresholds for removing clumbs - measured in cell sizes. Default: NULL}

\item{clump.directions}{directions for "clump-searching". Default: 8}

\item{env.rsaga}{environment of SAGA GIS. Default: RSAGA::rsaga.env()}

\item{CFR.buf}{buffer applied on contrast-filter-results, measured in cell-sizes. Default: "1"}

\item{output.path}{path for output files. Default: tempdir()}

\item{writeCFRaster}{write contrast-filter results. Default: FALSE}

\item{writeRaster.NAflag}{NoData values for NA values in raster. Default: -99999}

\item{freeMemory}{free memory after contrast filtering. Default: TRUE}

\item{show.output.on.console}{show output of tools on console. Default: FALSE}

\item{quiet}{do not show comments and process time in console. Default: TRUE}

\item{Grass.Segmentation.Threshold}{threshold used of merging segments. Default: 0.24}

\item{Grass.Segmentation.Minsize}{minsize of a segment. Default: 0}

\item{Grass.Segmentation.Memory}{memory used in computation. Default: 1024}

\item{Segments.Poly}{output location for segmented objects (shapefile). Default: paste0(tempdir(), "/", "outSegPoly.shp")}

\item{Segments.Grid}{output location for segmented grid. Default: paste0(tempdir(), "/", "outSegGrid.sgrd")}

\item{defaultGrass}{GRASS GIS environmet for temporal setup, see package link2GI. Default: c("C:/OSGeo4W64", "grass-7.2.2", "OSGeo4W64")}

\item{fill.holes}{polygon segmentation output contains holes. Default: FALSE}

\item{Fast.Representativeness.LevelOfGeneralisation}{level of generalisation for computing seed points. Default: "2.0"}
}
\value{
\linkS4class{SpatialPolygonsDataFrame} containing the segments/objects
}
\note{
\itemize{
  \item \emph{r} is stretched to gray-scale (0-255) during process by linear normalization
}
}
\keyword{EBImage,}
\keyword{SAGA}
\keyword{conversion,}
\keyword{gray-scale-image,}
\keyword{growing,}
\keyword{matrix,}
\keyword{raster,}
\keyword{seeded-region}
