% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convR2GSM.R
\name{convR2GSM}
\alias{convR2GSM}
\title{Conversion of raster layer to gray-scale image}
\usage{
convR2GSM(r, NA.val.in = 0)
}
\arguments{
\item{r}{\linkS4class{RasterLayer} to be converted either to an image or a matrix}

\item{NA.val.in}{replace value for no data (NA) in \emph{r}. Default: 0}
}
\value{
\linkS4class{data.table} containing object-based textural features
}
\description{
Conversion of raster layer to gray-scale image
}
\note{
\itemize{
  \item \emph{r} is stretched to gray-scale (0-255) during process by linear normalization
}
}
\keyword{EBImage,}
\keyword{conversion}
\keyword{gray-scale-image,}
\keyword{matrix,}
\keyword{raster,}
