% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctDBF.R
\name{correctDBF}
\alias{correctDBF}
\title{Correct .dbf of file}
\usage{
correctDBF(x, end.n = length(colnames(d$dbf)), adjust.n = 0,
  new.colnames)
}
\arguments{
\item{x}{full path to .dbf-file, but without extension}

\item{end.n}{start number of adjusting. Default: number of columns}

\item{adjust.n}{number of columns that are renamed. Default: 0}

\item{new.colnames}{vector with new names}
}
\description{
This function enables the possibility to overwrite the header of an existing .dbf-file.
Some applications, such as SAGA GIS, store shapefiles in formats that are not conform with the
ESRI-shapefile format. By this function, "corrupt" headers can be replaced.
}
\note{
length of vector with new names must be identical to number of adjusted columns
}
\keyword{correct}
\keyword{dbf}
\keyword{dbf,}
