% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effSurvArea.R
\name{effSurvArea}
\alias{effSurvArea}
\title{Calculation of effective surveyed area}
\usage{
effSurvArea(elev, pts, maxdist = 1000, do.extend = FALSE,
  path.save = tempdir(), path.temp = tempdir(), method = "bilinear",
  memory = 4096, NAflag = -99999, return.geom = TRUE, quiet = TRUE,
  show.output.on.console = FALSE)
}
\arguments{
\item{elev}{\linkS4class{RasterLayer} containing elevation values}

\item{pts}{\linkS4class{sf} containing the points from which the effective surveyed area is estimated.}

\item{maxdist}{viewshed distance to analysie. Default: 1000 m}

\item{do.extend}{extend output to extent of input elev. Default: FALSE}

\item{path.save}{path to store files. Default: tempdir()}

\item{path.temp}{path to store tempory files. Default: tempdir()}

\item{method}{method for extracting elevation values for points. Default: 'bilinear'. Use 'simple' for neares neighbor.}

\item{memory}{memory used in viewshed computation. GRASS GIS r.viewshed. Default: 4096.}

\item{NAflag}{replace value for NA (NULL data). Default: -99999.}

\item{return.geom}{If TRUE geometry is returned by functtion. Default: TRUE.}

\item{quiet}{do not show comments and state. Default: TRUE}

\item{show.output.on.console}{show GRASS GIS console output. Default: FALSE}
}
\value{
list of \linkS4class{RasterLayer} of viewshed parameters
}
\description{
This function calculate the effective surveyed area based on viewshed analysis.
A valid GRASS GIS session must be initiated before.
}
\note{
\itemize{
  \item function is taken from Bornaetxea, T., Rossi, M., Marchesini, I., & Alvioli, M. (2018). Effective surveyed area and its role in statistical landslide susceptibility assessments. Natural Hazards and Earth System Sciences, 18(9), 2455-2469.
}
}
\keyword{GIS,}
\keyword{GRASS}
\keyword{area,}
\keyword{effective}
\keyword{surveyed}
\keyword{viewshed}
