% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESP.R
\name{ESP}
\alias{ESP}
\title{ESP - ESTIMATION OF SCALE PARAMETER}
\usage{
ESP(Scale.Input.Grid, Scale.Input.Grid.Cell.Size = "1",
  Scale.Statistic.Min.Size = "0", ESP.save = FALSE, ESP.save.path = NULL,
  Count = "1", Min = "0", Max = "0", Range = "0", Sum = "0",
  Mean = "0", Var = "0", Stddev = "1", Quantile = 0, Scales,
  Grass.ESP.Method = "Threshold", ...)
}
\arguments{
\item{Scale.Input.Grid}{input grid for computing segmentation scale parameters}

\item{Scale.Input.Grid.Cell.Size}{cell size of input grid. Default: "1"}

\item{Scale.Statistic.Min.Size}{min size of area/polygon which is included in statistics (usefull for SAGA GIS segmentations). Default: "0"}

\item{ESP.save}{save estimations of function. Default: FALSE}

\item{ESP.save.path}{path where estimations are stored. Default: input path of \emph{segment.poly}}

\item{Count}{amount of cells (Grid Statistics). Default:"1"}

\item{Min}{minimum value (Grid Statistics). Default:"0"}

\item{Max}{maximum value (Grid Statistics). Default:"0"}

\item{Range}{range of values (Grid Statistics). Default:"0"}

\item{Sum}{sum of values (Grid Statistics). Default:"0"}

\item{Mean}{mean of values (Grid Statistics). Default:"1"}

\item{Var}{variance of values (Grid Statistics). Default:"0"}

\item{Stddev}{standard deviation (Grid Statistics). Default:"1"}

\item{Quantile}{qunatile (Grid Statistics). Default:"0"}

\item{Scales}{containing scale parameters for loop-segmentation}

\item{Grass.ESP.Method}{determining on which parameter the objective function (~loop) should be performed. Default: "Threshold"}

\item{other...}{see \code{\link{segmentation}}}
}
\description{
The ESP function by Dragut et al. (2010) calculates optimal
segmentation scales by calculating, at first, the local variance for every segmented images
and then by comparing the results of different segmentation levels (ROC-LV).
}
\note{
\itemize{
  \item DRAGUT, L., TIEDE, D. & S.R. LEVICK (2010): ESP: a tool to estimate
    scale parameter for multiresolution image segmentation of remotely
    sensed data. - International Journal of Geographical Information
    Science 24, 6, 859-871.
}
}
\examples{
Objective.Function()

}
\keyword{ESP,}
\keyword{USPO}
