% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRainEventData.R
\name{findRainEvent}
\alias{findRainEvent}
\title{findRainEvent}
\usage{
findRainEvent(x = x, x.pos.dryPeriods)
}
\arguments{
\item{x}{vector containing precipitation}

\item{x.pos.dryPeriods}{indices of x containing dry periods. Result of findRainFallPosition.}
}
\value{
list containing rain events as specific indices of x.
}
\description{
This function return a list of rain events.
}
