% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRainEventData.R
\name{findRainFallPosition}
\alias{findRainFallPosition}
\title{findRainFallPosition}
\usage{
findRainFallPosition(x, dates, rainThresh, rainOffLength, op.rainThresh,
  op.rainOffLength, index.month.warm.season)
}
\arguments{
\item{x}{vector containing precipitation}

\item{dates}{vector containing months of dates. The length of dates must be similar to the length of x}

\item{rainThresh}{list containing indices of rain events}

\item{rainOffLength}{modus of calculation of rain metrics. "sub" of sub-rainfall events or "main" for critical rainfall event}

\item{op.rainThresh}{operator for rain-threshold: x OP rainThresh}

\item{op.rainOffLength}{operator for rainOff-threshold: lengths of event OP rainOffLength-Threshold}
}
\value{
vector containing indices of x corresponding to specific selection.
}
\description{
This function return indices of a specific selection.
}
