% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genPntsByNet.R
\name{genPntsByNet}
\alias{genPntsByNet}
\title{Generate points based on distance constraints of a fish net}
\usage{
genPntsByNet(x, r, dist, do.PointsOnSurface = TRUE, quiet = TRUE,
  save.path = NULL)
}
\arguments{
\item{x}{object of class sf}

\item{r}{raster layer providing the extent of the fish net}

\item{dist}{distance between points}

\item{do.PointsOnSurface}{if TRUE than points on surface are generated for missed objects. Default: TRUE}

\item{quiet}{no outputs in console. Default: TRUE}

\item{save.path}{full path to store the result. Default: NULL}
}
\value{
list of 1. points of interest, 2. all points
}
\description{
This function generate points by a fish net (center points). The fish net is created
based on a input raster providing the bounding box and a given cell size.
}
\keyword{constraint,}
\keyword{distance}
\keyword{fish}
\keyword{net}
\keyword{points,}
\keyword{structured}
