% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genRandomPnts.R
\name{genRandomPnts}
\alias{genRandomPnts}
\title{Generate random points based on distance constraints}
\usage{
genRandomPnts(x, seed = 123, dist = NULL, n = Inf, maxit = 100,
  quiet = TRUE, cores = 1, ...)
}
\arguments{
\item{x}{object of class sf}

\item{seed}{seed parameter for replication. Default: 123}

\item{dist}{inhibition distance of spatstat::rSSI. Default: NULL}

\item{n}{maximum number of points allowed. If n is finite, stop when the total number of points in the point pattern reaches n. If n is infinite (the default), stop only when it is apparently impossible to add any more points}

\item{maxit}{number of rejected proposals after which the algorithm should terminate. Default: 100}

\item{quiet}{no outputs in console. Default: TRUE}

\item{cores}{number of cores for parallel processing. Default: 1 (sequential)}

\item{...}{additional parameters for spatstat::rSSI-function}
}
\value{
Randomly sampled points of class sf
}
\description{
This function generate random points inside polygons based on distance constraints.
The function is based on spatstat::rSSI.
}
\keyword{constraint}
\keyword{distance}
\keyword{points,}
\keyword{random}
