% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genRandomPntsByGrd.R
\name{genRandomPntsByGrd}
\alias{genRandomPntsByGrd}
\title{Generate random points based on distance constraints}
\usage{
genRandomPntsByGrd(x, r, field, env.grass, do.PointsOnSurface = TRUE, dist,
  save.path = NULL, quiet = TRUE, seed = 12345)
}
\arguments{
\item{x}{object of class sf}

\item{r}{rsater layer object}

\item{field}{field of x object for rasterisation}

\item{env.grass}{setting for GRASS GIS initiation}

\item{do.PointsOnSurface}{if TRUE than points on surface are generated for missed objects. Default: TRUE}

\item{dist}{minimum distance between points}

\item{save.path}{full path to store the result. Default: NULL}

\item{quiet}{no outputs in console. Default: TRUE}

\item{seed}{number to determine the randomness. Default: 12345}

\item{...}{additional parameters for spatstat::rSSI-function}
}
\value{
Randomly sampled points of class sf
}
\description{
This function generate random points inside polygons based on distance constraints.
The function is based on spatstat::rSSI.
}
\keyword{constraint}
\keyword{distance}
\keyword{points,}
\keyword{random}
