% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeomorphometryGradientMetricsToolbox.R
\name{GeomorphometryGradientMetricsToolbox}
\alias{GeomorphometryGradientMetricsToolbox}
\title{Geomorphometry & Gradient Metrics Toolbox}
\usage{
GeomorphometryGradientMetricsToolbox(elevation, slope = NULL, aspect = NULL,
  tool = c("Linear Aspect", "Landfrom", "Mean Slope", "SEI", "RI"),
  params = params, zScale = "1.0", output.path = tempdir(),
  quiet = TRUE)
}
\arguments{
\item{elevation}{\linkS4class{RasterLayer} with elevation data}

\item{slope}{\linkS4class{RasterLayer} with slope values. As default the slope will be computet using \emph{RQGIS::run_qgis(alg = "grass7:r.slope.aspect")}. Default: NULL}

\item{aspect}{\linkS4class{RasterLayer} with aspect values. As default the aspect will be computet using \emph{RQGIS::run_qgis(alg = "grass7:r.slope.aspect")}. Default: NULL}

\item{tool}{vector containg the name of tools. Actually, the following are supported: "Linear Aspect", "Landfrom", "Mean Slope",  "SEI", "RI". Default: c("Linear Aspect", "Landfrom", "Mean Slope",  "SEI", "RI")}

\item{params}{list of lists containing function parameter for functions in tool. There are basically three parameters defining the (1) "size" of moving window, (2) the filename, and (3) a logical value if the raster shall be written (writeRaster).  I.e. for "Linear Aspect" - list(size = 3, filename = "LinAspect.tif", writeRaster = TRUE). Default: see notice}

\item{zScale}{multiplicative factor to convert elevation units to horizontal units. Default: "1.0"}

\item{output.path}{path to output folder. Default: tempdir()}

\item{quiet}{no outputs in console. Default: TRUE}
}
\value{
List of \linkS4class{RasterLayer} containing computed geomorphometric indices or gradient models
}
\description{
This toolbox contains different function for calculating geomorphometric indices or gradient models.
The toolbox is based on the Geomorphometry & Gradient Metrics Toolbox from Jeffrey Evans.
Actually, it supports the computation of 'Linear Aspect', 'Mean Slope', 'Site Exposure Index',
'Landform' (concavity/convexity landform index (Bolstads variant)) and 'Roughness Index'.
}
\note{
\itemize{
  \item GRASS GIS algorithm (grass7:r.slope.aspect) is similar to ArcGIS slope/aspect
  \item default for params: list(list(size = 3, filename = "LinAspect.tif", writeRaster = TRUE), list(size = 3, filename = "Landform.tif", writeRaster = TRUE), list(size = 3, filename = "MeanSlp.tif", writeRaster = TRUE), list(size = NA, filename = "SEI.tif", writeRaster = TRUE), list(size = 3, filename.RI = "RI.tif", filename.RIw = "RIw.tif", writeRaster = TRUE))
}
}
\keyword{&}
\keyword{Aspect,}
\keyword{Evans,}
\keyword{Exposure}
\keyword{Geomorphometry}
\keyword{Gradient}
\keyword{Index}
\keyword{Index,}
\keyword{Jeffrey}
\keyword{Landform,}
\keyword{Linear}
\keyword{Mean}
\keyword{Metrics}
\keyword{Roughness}
\keyword{Site}
\keyword{Slope,}
\keyword{Toolbox,}
