% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBoundingBox.R
\name{getBoundingBox}
\alias{getBoundingBox}
\title{Rotated and scaled bounding box}
\usage{
getBoundingBox(spdf, col.name, scale.factor = 2, k = 2,
  projection = NA, centroid = FALSE, set.centroid = "inverse",
  k.centroid = 2, scale.side = "long", cores = 1, quiet = TRUE)
}
\arguments{
\item{spdf}{SpatialPolygonsDataFrame. SPDF must contains a valid \emph{ID} field and a field containing direction angles}

\item{col.name}{name or number of column containing angles in degree}

\item{scale.factor}{scale factor for bounding box. This can also be vector with two elements: 1. for first side (default: small), and 2. second side (default: longside). Default: 2}

\item{k}{factor controlling the distance between first two coordinates related to the side and centroid. Default: 2}

\item{projection}{projection of \emph{spdf} in CRS-format. If \emph{spdf} is projected, projection is read from \emph{spdf}. Default: NA}

\item{centroid}{centroid is taken for the bounding box. Default: FALSE}

\item{set.centroid}{set intial starting point of bounding box with respect to the centroid. Default: "inverse" (alternative: "direction")}

\item{scale.side}{side which is scaled in direction. Default: "long" (alternative: small)}

\item{cores}{number of cores for parallel processing. Default: 1 (sequential)}

\item{quiet}{show output on console. Default: FALSE}

\item{k.centroid.flow}{same as \emph{k}, used when \emph{centroid} is set to FALSE}
}
\description{
This function creates a bounding box for every SpatialPolygonsDataFrame object.
Moreover, the bounding box is rotated and scaled in a direction given by the user. \cr \cr
}
\details{
The calculation is performed by trigonometric functions. \cr
R takes angles in the following anti-clockwise sense: \cr EAST 0°, NORTH 90°, WEST 180°, SOUTH 270°. \cr \cr
However, the angles in the output have the following clockwise sense: \cr
NORTH 0°, EAST 90°, SOUTH 180°, WEST 270°.
}
\note{
\itemize{
 \item SpatialPolygonsDataFrame must have a valid \emph{ID} field and a column specifying
        the angles in which the bounding box is rotated
 \item Angles will be transformed to clockwise sense: NORTH 0°, EAST 90°, SOUTH 180°, WEST 270°.
 }
}
\keyword{bounding}
\keyword{box,}
\keyword{direction}
\keyword{rotated,}
\keyword{scaled,}
