% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRainThreshFreqM.R
\name{getRainThreshFreqM}
\alias{getRainThreshFreqM}
\title{Get rain event of a landslide}
\usage{
getRainThreshFreqM(Re, D, method = "NLS", prob.threshold = 0.05,
  log10.transform = FALSE, bootstrapping = TRUE, R = 1000,
  seed = 123, use.integralError = TRUE, ...)
}
\arguments{
\item{Re}{vector containing the rain event variable, e.g. cumulated event rainfall (in mm) or intensity (mm/h)}

\item{D}{vector containing the duration of the rainfall events}

\item{method}{method to compute threshold. Either "LS" for least square, or "NLS" for non-linear least squares Method. Default: "nls"}

\item{prob.threshold}{exceedance probability level. Default: 0.05 (5 [percent] )}

\item{log10.transform}{log-transformation of input vectors Re and/or D. Default: TRUE}

\item{bootstrapping}{If TRUE bootstrapping is performed. Default: TRUE}

\item{R}{the number of bootstrap replicates, see boot::boot() for more information. Default: 1000}

\item{seed}{replicable bootstrapping. Default: 123}

\item{use.integralError}{for estimating x of prob.threshold, the entire function is integrated first to estimate the bias: (1 - INTEGRAL)/2. Default. TRUE}

\item{...}{more options passed to boot:boot() function, i.e. parallel for paralell processing}
}
\value{
vector containing rainfall metrics (see description). If return.DataFrame is TRUE a data.frame is returned containing similar rain metrics for all rain events.
}
\description{
This function calcuates different precipitation characteristics for a specific time-series:
total precipitation, number of rainfall events, weighted mean intensitiy of rainfall events (normalized by MAP, RD or RDN),
cumulative critical event rainfall (normalized by MAP, RD or RDN), maximum rainfall during critical rainfall event,
duration of critical rainfall event, critical rainfall intensitiy (normalized by MAP, RD or RDN), rainfall at day of failure (start date),
rainfall intensity at day of failure (start date), maximum rainfall at day of failure (start date).
}
\note{
\itemize{
  \item Brunetti, M. T., Peruccacci, S., Rossi, M., Luciani, S., Valigi, D., & Guzzetti, F. (2010). Rainfall thresholds for the possible occurrence of landslides in Italy. Natural Hazards and Earth System Sciences, 10(3), 447.
  \item Peruccacci, S., Brunetti, M. T., Luciani, S., Vennari, C., & Guzzetti, F. (2012). Lithological and seasonal control on rainfall thresholds for the possible initiation of landslides in central Italy. Geomorphology, 139, 79-90.
  \item Rossi, M., Luciani, S., Valigi, D., Kirschbaum, D., Brunetti, M. T., Peruccacci, S., & Guzzetti, F. (2017). Statistical approaches for the definition of landslide rainfall thresholds and their uncertainty using rain gauge and satellite data. Geomorphology, 285, 16-27.
  \item Guzzetti, F., Peruccacci, S., Rossi, M., & Stark, C. P. (2007). Rainfall thresholds for the initiation of landslides in central and southern Europe. Meteorology and atmospheric physics, 98(3-4), 239-267.
}
}
\keyword{appraoch}
\keyword{automatic}
\keyword{event,}
\keyword{landslide,}
\keyword{rainfall}
\keyword{tresholds,}
