% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiPassSegmentation.R
\name{hiPassSegmentation}
\alias{hiPassSegmentation}
\title{Contrast-Filter-Segmentation}
\usage{
hiPassSegmentation(input.filter, input.segmentation = input.filter,
  result.filter = NULL, scale.factor, threshold,
  env.rsaga = RSAGA::rsaga.env(), writeRaster.NAflag = -99999,
  show.output.on.console = FALSE, quiet = TRUE,
  Grass.Segmentation.Threshold = 0.24, Grass.Segmentation.Minsize = 1,
  Grass.Segmentation.Memory = 1024,
  Segments.Poly = file.path(tempdir(), "outSegPoly.shp"),
  Segments.Grid = file.path(tempdir(), "outSegGrid.sgrd"),
  load.output = FALSE, fill.holes = FALSE,
  estimateScaleParameter = FALSE, ...)
}
\arguments{
\item{input.filter}{\linkS4class{RasterLayer} on which the filtering is applied}

\item{input.segmentation}{\linkS4class{RasterLayer} for segmentation. Default: input.filter}

\item{env.rsaga}{environment of SAGA GIS. Default: RSAGA::rsaga.env()}

\item{writeRaster.NAflag}{NoData values for NA values in raster. Default: -99999}

\item{show.output.on.console}{show output of tools on console. Default: FALSE}

\item{quiet}{do not show comments and process time in console. Default: TRUE}

\item{Grass.Segmentation.Threshold}{threshold used of merging segments. Default: 0.24}

\item{Grass.Segmentation.Minsize}{minsize of a segment. Default: 0}

\item{Grass.Segmentation.Memory}{memory used in computation. Default: 1024}

\item{Segments.Poly}{output location for segmented objects (shapefile). Default: paste0(tempdir(), "/", "outSegPoly.shp")}

\item{Segments.Grid}{output location for segmented grid. Default: paste0(tempdir(), "/", "outSegGrid.sgrd")}

\item{estimateScaleParameter}{must be only be TRUE when scale parameter function is used. Default: FALSE}

\item{...}{parameters for function hiPassThresh}

\item{Fast.Representativeness.LevelOfGeneralisation}{level of generalisation for computing seed points. Default: "2.0"}
}
\value{
\linkS4class{SpatialPolygonsDataFrame} containing the segments/objects
}
\description{
This function performs a contrast filter segmentation on an input \linkS4class{RasterLayer}.
In the firest stept the raster is converted to a gray-scale matrix, followed by filter functions of the EBImage package.
Then, the filtering is applied as mask on the segmentation-input. Afterwards, a segmentation is performed
using GRASS GIS region growing. Finally, the segments are transformed to \linkS4class{SpatialPolygonsDataFrame}
and returned as function output.
In the future this funcion will be included in the segmentation-function.
}
\keyword{GIS}
\keyword{GIS,}
\keyword{GRASS}
\keyword{SAGA}
\keyword{conversion,}
\keyword{filtering,}
\keyword{growing,}
\keyword{high-pass}
\keyword{matrix,}
\keyword{raster,}
\keyword{seeded-region}
