% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiPassThresh.R
\name{hiPassThresh}
\alias{hiPassThresh}
\title{Using high-pass filter and thresholds to filter image}
\usage{
hiPassThresh(x, scale.factor, threshold, path.output = NULL,
  do.sieve = TRUE, sieve.mode = "0", sieve.thresh = 4,
  do.shrink.expand = TRUE, expand = 4, do.buf = FALSE,
  buf.size = 1, do.use.temp.HPF = FALSE, do.use.temp.Thresh = FALSE,
  path.save = tempdir(), NoData = -99999, env.rsaga = NULL,
  show.output.on.console = FALSE, quiet = TRUE, ...)
}
\arguments{
\item{x}{raster object or input path of file}

\item{path.output}{= NULL}

\item{do.sieve}{= TRUE}

\item{sieve.mode}{mode of cell-neighborhood,  Neumann [0]: the four horizontally and vertically neighboured cells; Moore [1]: all eight adjacent cells. Default: "0"}

\item{sieve.thresh}{minimum number of cells in a group of adjacent cells (minimum 2). Default: 4}

\item{do.shrink.expand}{regions with valid data in the input grid can be shrinked or expanded by a certain amount (radius). Default: TRUE}

\item{expand}{maximum expansion distance (radius). Default: 4 [cells]}

\item{do.buf}{= FALSE}

\item{buf.size}{buffer size in cell's value. Default: 1}

\item{do.use.temp.HPF}{already existing high-pass filter files are used. Default: FALSE}

\item{do.use.temp.Thresh}{already existing thresholding files are used. Default: FALSE}

\item{path.save}{output path for saving objects. Default: tempdir()}

\item{NoData}{no data value. Default: -99999}

\item{env.rsaga}{environment of SAGA GIS. As default, the environment will be automatically estimated. Default: NULL}

\item{show.output.on.console}{show output on console. Default: FALSE}

\item{quiet}{no outputs in console. Default: TRUE}
}
\value{
raster object
}
\description{
This function performs a high-pass filter followed by a thresholding procedure in order to
select specific cluster of cells.
}
\note{
\itemize{
  \item SAGA GIS must be installed.

}
}
\keyword{GIS}
\keyword{SAGA}
\keyword{filter,}
\keyword{high-pass}
