% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initGRASSWin.R
\name{initGRASSWin}
\alias{initGRASSWin}
\title{Initialisation of rgrass7 in Windows}
\usage{
initGRASSWin(x = NULL, osgeo4w.root = "C:\\\\OSGEO4W64",
  grass.version = "grass-7.2.2",
  initGRASS.path = "C:/OSGeo4W64/apps/grass/grass-7.2.2", set.SysEnv = TRUE,
  use.link2GI = FALSE, link2GI.defaultGrass = c("C:/OSGeo4W64",
  "grass-7.2.2", "OSGeo4W64"), quiet = TRUE)
}
\arguments{
\item{x}{\linkS4class{RasterLayer}  for region and projection. Default: NULL}

\item{osgeo4w.root}{root of OSGeo installation. Default: "C:\\OSGEO4W64"}

\item{grass.version}{GRASS GIS version. Default: "grass-7.2.2"}

\item{initGRASS.path}{initiation path of GRASS GIS through rgrass7. Default: "C:/OSGeo4W64/apps/grass/grass-7.2.2"}

\item{set.SysEnv}{set windows environmental path in R session. Default: TRUE. If FALSE no path is set.}

\item{use.link2GI}{use of link2GI package of initialisation. Default: FALSE}

\item{link2GI.defaultGrass}{initialisation using link2GI package. Default: c("C:/OSGeo4W64", "grass-7.2.2", "OSGeo4W64"))}

\item{quiet}{show output on console. Default: TRUE}
}
\description{
This funcion helps to initialise rgrass7 using windows.
}
\keyword{rgrass7,}
\keyword{windows}
