% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kMeansThresholding.R
\name{kMeansThresholding}
\alias{kMeansThresholding}
\title{Thresholds by k-Means clustering}
\usage{
kMeansThresholding(df, n.clust = NULL, G = 25, size.data = NULL,
  size.sample = 5000, seed = 123, iter.max = 5000)
}
\arguments{
\item{df}{input data frame}

\item{n.clust}{number of initial cluster (Default: NULL)}

\item{G}{see \code{\link[mclust]{Mclust}}. Default: 25}

\item{size.data}{percentage of input data used in anaylsis. Default: NULL}

\item{size.sample}{sample in initialization of \code{\link[mclust]{Mclust}} in order to speed up the process. Default: 5000, to use full data set to NULL}

\item{seed}{set seed. Default: 123}

\item{iter.max}{number of iterlations in \code{\link[mclust]{Mclust}}. Default: 5000}
}
\value{
vector containing the cluster centers
}
\description{
This function calculates thresholds by using k-Means clustering. The number of clusters is automatically derived by the Schwarz Bayesian criterion
}
\note{
\itemize{
  \item see \href{https://cran.r-project.org/web/packages/mclust/mclust.pdf}{mclust} \emph{(last call: 13-04-2017)}
  \item see \href{https://stat.ethz.ch/R-manual/R-devel/library/stats/html/kmeans.html}{kmeans} \emph{(last call: 13-04-2017)}
}
}
\examples{
kMeansThresholding()

}
\keyword{Bayesian}
\keyword{Schwarz}
\keyword{clustering,}
\keyword{criterio}
\keyword{k-Means}
\keyword{thresholding,}
