% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lengthWidthRatio.R
\name{lengthWidthRatio}
\alias{lengthWidthRatio}
\title{Length-Width-Ratio}
\usage{
lengthWidthRatio(spdf, cores = 1, quiet = TRUE)
}
\arguments{
\item{spdf}{SpatialPolygonsDataFrame input}

\item{cores}{number of cores for parallel processing. Default: 1 (sequential)}

\item{quiet}{no outputs in console. Default: TRUE}
}
\value{
data.table containing the \emph{ID} field and a column (\emph{ratio}) with length-width-ratio
}
\description{
This function calculates the length-width-ratio for SpatialPolygonsDataFrame objects.
The length-width vectors for the ratio are computed by a principal component analysis according to YI & MARSHALL (2000)
}
\note{
\itemize{
  \item for the principal component analysis the \code{\link[stats]{princomp}} function is used
  \item SpatialPolygonsDataFrame input must have an \emph{ID} field with unique numbers
  \item YI, W.& S. MARSHALL (2000): Principal component analysis in application
        to object orientation. Geo-spatial Information Science, 3(3), 76-78.
}
}
\keyword{analysis}
\keyword{component}
\keyword{length-width-ratio,}
\keyword{principal}
