% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainDirection.R
\name{mainDirection}
\alias{mainDirection}
\title{Object orientation or main direction}
\usage{
mainDirection(spdf, angle = "main", quiet = TRUE)
}
\arguments{
\item{spdf}{\linkS4class{SpatialPolygonsDataFrame} input}

\item{angle}{angle taken for output. Default: "main" (alternative: "weighted main")}

\item{quiet}{no outputs in console. Default: TRUE}
}
\value{
data.table containing the \emph{ID} field, a column (\emph{angle}) containing the object orientations,
 and a column (\emph{angle_inv}) contaning the inverse object orientations. Both object orientations are necessairy
 due to the ambiguity of the orientation in xy-space without fix starting point
}
\description{
This function calculates the object orientation (or main direction) for SpatialPolygonsDataFrame objects
in xy-space. For this, a principal component analysis is performed according to YI & MARSHALL (2000)
}
\note{
\itemize{
  \item for the principal component analysis the \code{\link[stats]{princomp}} function is used
  \item \linkS4class{SpatialPolygonsDataFrame} input must have an \emph{ID} field with unique numbers
  \item YI, W.& S. MARSHALL (2000): Principal component analysis in application
        to object orientation. Geo-spatial Information Science, 3(3), 76-78.
}
}
\keyword{analysis}
\keyword{component}
\keyword{direction,}
\keyword{main}
\keyword{object}
\keyword{orientation,}
\keyword{principal}
