% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborDirection.R
\name{neighborDirection}
\alias{neighborDirection}
\title{Calculation of the direction of every object to its neighbors}
\usage{
neighborDirection(spdf, col.name, modus = "nb", tol = 360, spdf.bb = NULL,
  sp.PoS = NULL, nb = NULL, bb = NULL, quiet = TRUE, ...)
}
\arguments{
\item{spdf}{\linkS4class{SpatialPolygonsDataFrame}, or full path of a shapefile}

\item{col.name}{column name for angles (in degree, N = 0, E = 90, S = 180, W = 270)}

\item{modus}{"bb" for using a boundig box or "nb" for using object neighborhood \emph{nb}. Modus \emph{bb} returns data.frame based on spdf. Modus \emph{nb} returns list based on nb.}

\item{tol}{tolerance for select neighbor of specific range. Default: 360}

\item{spdf.bb}{\linkS4class{SpatialPolygonsDataFrame} from which the bounding boxes are created}

\item{sp.PoS}{\linkS4class{SpatialPoints} to which the direction is calculated. As standard they are calculated from spdf input. Default: NULL}

\item{nb}{object neighborhood based on \code{\link[spdep]{poly2nb}}, for modus "nb. Default: Null}

\item{bb}{bounding box for modus "bb". Default: Null}

\item{quiet}{no outputs in console. Default: TRUE}
}
\value{
In the case of modus \emph{nb} a list is returned. If modus \emph{bb} is set, a data.frame based on the input spdf is returned
}
\description{
This function neighborDirection calculate the direction in degree to its surrounding neighbors or a selection
}
\note{
\itemize{
  \item \linkS4class{SpatialPolygonsDataFrame} input must have an \emph{ID} field with unique numbers
}
}
\keyword{analysis}
\keyword{direction,}
\keyword{image}
\keyword{neighbor}
\keyword{object-oriented}
