% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborGrowing.R
\name{neighborGrowing}
\alias{neighborGrowing}
\title{Growing of neighbors}
\usage{
neighborGrowing(spdf, nb, ID.start = NULL, ID.class = NULL,
  return.input = TRUE, return.gUnaryUnionNeighbors = TRUE, quiet = TRUE)
}
\arguments{
\item{spdf}{\linkS4class{SpatialPolygonsDataFrame}, or full path of a shapefile}

\item{nb}{object neighborhood based on \code{\link[spdep]{poly2nb}}}

\item{ID.start}{vector containing \emph{ID}'s where the growing starts. Default: NULL}

\item{ID.class}{\emph{ID}'s of a specific class in order to allow growing only for class members. Default: NULL}

\item{return.input}{\emph{spdf} is returned with neighbor information. Default: TRUE}

\item{return.gUnaryUnionNeighbors}{return united neighbors based on \code{\link[rgeos]{gUnaryUnion}}. Default: TRUE}

\item{quiet}{no outputs in console. Default: TRUE}
}
\value{
list containing input \linkS4class{SpatialPolygonsDataFrame} with information on growing neighbors, geometry with united neighbors, or both (default)
}
\description{
This function grows neighbor \linkS4class{SpatialPolygonsDataFrame} objects together
}
\note{
\itemize{
  \item \linkS4class{SpatialPolygonsDataFrame} input must have an \emph{ID} field with unique numbers
}
}
\examples{
neighborGrowing()

}
\keyword{analysis}
\keyword{growing,}
\keyword{image}
\keyword{neighbor}
\keyword{object-oriented}
\keyword{spdep,}
