% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectFeatures.R
\name{objectFeatures}
\alias{objectFeatures}
\title{Caclulation of object features}
\usage{
objectFeatures(r.seg, quiet = TRUE)
}
\arguments{
\item{r.seg}{segments or objects in form of a \linkS4class{RasterLayer} (only 16 or 32 bit integer)}

\item{quiet}{no outputs in console. Default: TRUE}
}
\value{
\linkS4class{data.table} containing object features
}
\description{
This function calculates object features based on the \code{\link[EBImage]{computeFeatures}} function
}
\note{
\itemize{
  \item as input only 16 or 32 bit integer values are allowed
  \item see \code{\link[EBImage]{computeFeatures}}
}
}
\keyword{EBImage,}
\keyword{computeFeatures,}
\keyword{segmentation}
