% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObjectiveFunction.R
\name{Objective.Function}
\alias{Objective.Function}
\title{Objective Function}
\usage{
Objective.Function(Scale.Input.Grid, Scale.Input.Grid.Cell.Size = "1",
  Scale.Statistic.Min.Size = "0", Objective.Function.save = FALSE,
  Objective.Function.save.path = NULL, Count = "1", Min = "0",
  Max = "0", Range = "0", Sum = "0", Mean = "1", Var = "1",
  Stddev = "0", Objective.Function.save.name = "",
  Objective.Function.Mean.Segmentation.Grid = paste0(tmp.path,
  "Objective.Function.Mean.Segmentation.Grid.sgrd"),
  Objective.Function.Count.Grid = paste0(tmp.path,
  "Objective.Function.Count.Grid.sgrd"),
  Objective.Function.MoransI = paste0(tmp.path,
  "Objective.Function.MoransI.txt"), Quantile = 0, Scales,
  Grass.Objective.Function.Method = "Threshold", ...)
}
\arguments{
\item{Scale.Input.Grid}{input grid for computing segmentation scale parameters}

\item{Scale.Input.Grid.Cell.Size}{cell size of input grid. Default: "1"}

\item{Scale.Statistic.Min.Size}{min size of area/polygon which is included in statistics (usefull for SAGA GIS segmentations). Default: "0"}

\item{Objective.Function.save}{save estimations of function. Default: FALSE}

\item{Objective.Function.save.path}{path where estimations are stored. Default: input path of \emph{segment.poly}}

\item{Count}{amount of cells (Grid Statistics). Default:"1"}

\item{Min}{minimum value (Grid Statistics). Default:"0"}

\item{Max}{maximum value (Grid Statistics). Default:"0"}

\item{Range}{range of values (Grid Statistics). Default:"0"}

\item{Sum}{sum of values (Grid Statistics). Default:"0"}

\item{Mean}{mean of values (Grid Statistics). Default:"1"}

\item{Var}{variance of values (Grid Statistics). Default:"1"}

\item{Stddev}{standard deviation (Grid Statistics). Default:"0"}

\item{Objective.Function.save.name}{name for file. Default: ""}

\item{Objective.Function.Mean.Segmentation.Grid}{- output grid with mean values for segments. Default: temp}

\item{Objective.Function.Count.Grid}{output grid with amount of cells in segments. Default: temp}

\item{Objective.Function.MoransI}{output text file with Morans'I. Default: temp}

\item{Quantile}{qunatile (Grid Statistics). Default:"0"}

\item{Scales}{containing scale parameters for loop-segmentation}

\item{Grass.Objective.Function.Method}{determining on what parameter the objective function (~loop) is performed. Default: "Threshold"}

\item{other...}{see \code{\link{segmentation}}}
}
\description{
The Objective Function by ESPINDOLA et al. (2006) finds optimal
scale parameters by the comparision of the sum of the normalized
local variance and the normalized Moran's I of different
segmented images.
}
\note{
\itemize{
  \item ESPINDOLA, G.M., CAMARA, G., REIS, I.A., BINS, L.S. & A.M. MONTEIRO (2006):
        Parameter selection for region-growing image segmentation algorithms using spatial autocorrelation.
         - International Journal of Remote Sensing 27, 14, 3035-3040.
}
}
\examples{
Objective.Function()

}
\keyword{USPO}
\keyword{function,}
\keyword{objective}
