% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectTextureHaralick.R
\name{objectTextureHaralick}
\alias{objectTextureHaralick}
\title{Caclulation of object-based textural features}
\usage{
objectTextureHaralick(r.texture, r.seg, texture.haralick.nbins = 32,
  texture.haralick.scales = c(1, 2), NA.val.in = 0, quiet = TRUE)
}
\arguments{
\item{r.texture}{\linkS4class{RasterLayer} for which the textural features are calculated}

\item{r.seg}{segments or objects in form of a \linkS4class{RasterLayer} (only 16 or 32 bit integer)}

\item{texture.haralick.nbins}{size of gray level co-occurrence matrix. Default: 32}

\item{texture.haralick.scales}{definition of scales on which the texture is measured. Default: c(1,2)}

\item{NA.val.in}{replace value for no data (NA) in \emph{r.texture}. Default: 0}

\item{quiet}{no outputs in console. Default: TRUE}
}
\value{
\linkS4class{data.table} containing object-based textural features
}
\description{
This function calculates textural features for objects by considering the values inside of the objects.
The calculation is based on the \code{\link[EBImage]{computeFeatures.haralick}} function
}
\note{
\itemize{
  \item \emph{r.texture} is stretched to gray-scale (0-255) during process by linear normalization
  \item \emph{r.seg} must consist either of 16 bit or of 32 bit integer values
  \item the calculated Haralick features are rotation invariant (average over all four possible directions)
  \item a smaller texture.haralick.nbins can give more accurate estimates of the correlation, because the number of events per bin is higher.
  While a higher value will give more sensitivity
  \item see \code{\link[EBImage]{computeFeatures.haralick}}
  \item see \href{http://earlglynn.github.io/RNotes/package/EBImage/Haralick-Textural-Features.html}{here} for more details on Haralick textural features \emph{(last call: 13-04-2017)}
}
}
\keyword{EBImage,}
\keyword{computeFeatures.haralick,}
\keyword{features}
\keyword{object-based}
\keyword{textural}
