% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optHiPassThresh.R
\name{optHiPassThresh}
\alias{optHiPassThresh}
\title{Find optimal hyper-parameters for function highPassThresholding}
\usage{
optHiPassThresh(x, inventory, range.scale.factor, range.threshold,
  path.runfile = NULL, env.rsaga, path.save = tempdir(),
  NoData = -99999, ..., cores = 1, quiet = TRUE)
}
\arguments{
\item{x}{raster object or input path of file}

\item{inventory}{binarized raster inventory of same extent and resolution as x}

\item{range.scale.factor}{range of different scale factors}

\item{range.threshold}{range of different thresholds}

\item{path.runfile}{full path for run file with .txt as ending. Default: NULL}

\item{env.rsaga}{SAGA GIS environment}

\item{NoData}{= -99999}

\item{...}{see highPassThresholding()}

\item{cores}{number of cores for parallel processing. Default: 1 (sequential)}

\item{quiet}{no outputs in console. Default: TRUE}
}
\value{
data.frame with clssification measurements ().
}
\description{
This functions optimize the hyper-parameter of the function highPassThresholding using standard
classification error measurements. As comparison a valid inventory in form of a binary classification
must be given.
}
\note{
\itemize{
  \item SAGA GIS must be installed.

}
}
\keyword{GIS}
\keyword{SAGA}
\keyword{filter,}
\keyword{high-pass}
