% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotObjectiveFunction.R
\name{plotObjectiveFunction}
\alias{plotObjectiveFunction}
\title{Plot result of Objective Function}
\usage{
plotObjectiveFunction(x, legend.position = "bottom", title = "",
  selected.Scale = NULL, selected.Scale.label = "selected scale",
  col.OF = "#DC143C", col.PF = "black", col.nIV = "#6495ED",
  col.nMI = "#8B4513", col.sSl = "gold", selected.Scale.size = 4,
  selected.Scale.shape = 1, selected.Scale.stroke = 1.2, ...)
}
\arguments{
\item{x}{result of function \code{\link{ObjectiveFunction}}}

\item{legend.position}{position of legend in plot. Default: "bottom"}

\item{title}{plot title. Default: ""}

\item{selected.Scale}{scale parameter for highlighting. Default: NULL}

\item{selected.Scale.label}{legend label. Default: "selected scale"}

\item{col.OF}{line color for objective function. Default: "#DC143C"}

\item{col.PF}{line color for plateau function. Default: "black"}

\item{col.nIV}{line color for normalized intrasegment variance. Default: "#6495ED"}

\item{col.nMI}{line color for normalized Moran's I. Default: "#8B4513"}

\item{col.sSl}{point color for selected scale parameter. Default: "gold"}

\item{selected.Scale.size}{scale size of selected point. Default: 4}

\item{selected.Scale.shape}{shape of selected point. Default: 1}

\item{selected.Scale.stroke}{width of outline of selected point. Default: 1.2}

\item{...}{ggplot-theme options}
}
\value{
\code{\link[ggplot2]{ggplot}} object
}
\description{
This function plots the result of  \code{\link{ObjectiveFunction}} function using
\code{\link[ggplot2]{ggplot}}
}
\keyword{Function,}
\keyword{Objective}
\keyword{ggplot2,}
\keyword{plot}
