% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relationalClassFunction.R
\name{relationalClassFunction}
\alias{relationalClassFunction}
\title{Calculation of object relations to a class}
\usage{
relationalClassFunction(spdf, nb, class.var, class.var2 = NULL, var,
  var2 = NULL, var.is.angle = FALSE, class.as.neighbors = FALSE,
  only.neighbors = TRUE, centroid.for.distance = TRUE,
  calc.nb.flow = FALSE, col.flow = NULL, bb.intersection = "rgeos",
  return.bb.flow = FALSE, scale.factor.flow = c(1, 1), bb = NULL,
  centroid.flow = FALSE, set.centroid.flow = "inverse",
  k.centroid.flow = 2, k.flow = 2, scale.side.flow = "small",
  quiet = TRUE)
}
\arguments{
\item{spdf}{\linkS4class{SpatialPolygonsDataFrame}, or full path of a shapefile}

\item{nb}{object neighborhood based on \code{\link[spdep]{poly2nb}}. Otherwise, neighborhood is computed by spdep::poly2nb(..., Queens = TRUE)}

\item{class.var}{defination of class. Can be string or integer, or an expression in form of a list. For examples: class = as.integer(1), class = list("> 2.5", "expression"), class = list(c("> 2", "&", "< 7"), "expression")}

\item{class.var2}{defination of a second class, see \emph{class.var}. Enables class-to-class relationship. Default: NULL}

\item{var}{vector with number or name defining 1. field of evaluation (i.e. slope) and 2. field of weights (i.e. area). For example: c("Slope", "Area") or c(1,2)}

\item{var.is.angle}{variable for statistc represents angles. Other mathematical computation. Default: FALSE}

\item{class.as.neighbors}{class object is also neighbor object. Default: FALSE}

\item{only.neighbors}{object statistics are computed only for neighbors. Default: TRUE (FALSE means computation for the entiry data)}

\item{calc.nb.flow}{include statistics for neighbors in flow direction. Default: FALSE}

\item{col.flow}{column specifying direction. Default: NULL}

\item{bb.intersection}{method for intersection with bounding box. Default: "rgeos" (alternative: "sp")}

\item{return.bb.flow}{return geoemtry of counding box. Default: FALSE}

\item{bb}{input of bounding box geometry. Default: NULL}

\item{quiet}{no outputs in console. Default: TRUE}

\item{other...}{see \code{\link{getBoundingBox}}}
}
\value{
\linkS4class{data.frame} containing object statistics, and optional the created geometry of \code{\link{getBoundingBox}}
}
\description{
This function calculates statistics from neighbors of a given class to the given class (input is \linkS4class{SpatialPolygonsDataFrame}).
Moreover, it is possible to calculate statistics for neighbors in a specific direction of a class object by setting
the parameter for the \emph{getBoundingBox} in the function call.
In addition, if the neighbors are attributed to another second class, also a class-to-class relationship can be performed.
Furthermore, attributing the first and the second class must not been necessarily hard coded, instead it is also possible to use expressions for the class-to-class relationship.
However, here it can happen that some objects are assigned to both classes. As default, these objects are excluded from the statistic, but they can be included by setting the specific argument \emph{(class.as.neighbors)}.
Moreover, the function is also capable to compute statistics on angle inputs by setting the function argument of \emph{var.is.angle} to TRUE
}
\note{
\itemize{
  \item object statistics to class according to ECOGNITION DEVELOPER (2014: 253-255, 375-381):
  \itemize{
      \item see \code{\link{relationalObjectFunction}}
      \item \emph{ang_m_cl} - mean angle of an object in relation to its surrounding class objects
      \item \emph{ang_d_cl} - mean difference angle of an object in relation to its surrounding class objects
      \item \emph{ang_d_cl_a} - absolute mean difference angle
      \item \emph{bor_cl_a} - absolute border of an object shared with neighboring objects of a defined class
      \item \emph{bor_cl_rel} - relative border of an object shared with neighboring objects of a defined class
      \item \emph{dist_cl} - distance of a not-neighbor object of a defined class to a defined class object
      \item \emph{flow_nb_cl} - object is located in flow direction (TRUE/FALSE)
  }
  \item ECOGNITION DEVELOPER (2014) Reference Book. Trimble Documentation, München, Germany
  \item see \code{\link{getBoundingBox}}
  \item \linkS4class{SpatialPolygonsDataFrame} objects with NA values are ignored
  \item \linkS4class{SpatialPolygonsDataFrame} objects without neighbors are ignored
  \item \linkS4class{SpatialPolygonsDataFrame} input must have a valid \emph{Class} and \emph{ID} field.
     \emph{ID} field must be contained of unique numbers
}
}
\keyword{a}
\keyword{analysis}
\keyword{class,}
\keyword{image}
\keyword{object-oriented}
\keyword{objects}
\keyword{of}
\keyword{relation}
\keyword{spdep,}
\keyword{to}
