% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceInvalids.R
\name{replaceInvalids}
\alias{replaceInvalids}
\title{Replace invalid fields in a data.frame or data.table}
\usage{
replaceInvalids(x, replace.value = -9999)
}
\arguments{
\item{x}{input data.frame or data.table}

\item{replace.value}{value that is used instead of NA, NAN, or NULL. Default: -9999}
}
\description{
This function replaces invalid fields (NA, NAN, and NULL) in a data.frame and data.table. These fields can lead
to erroneous attribute tables in desktop GIS software.
}
\note{
\itemize{
  \item see \href{http://stackoverflow.com/questions/7235657/fastest-way-to-replace-nas-in-a-large-data-table}{Fastest way to replace NAs in a large data.table} \emph{(last call: 13-04-2017)}
}
}
\examples{
replaceInvalids()

}
\keyword{NA,}
\keyword{NAN,}
\keyword{NULL,}
\keyword{data.frame,}
\keyword{data.table,}
\keyword{replace}
