% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skyViewFactor.R
\name{skyViewFactor}
\alias{skyViewFactor}
\title{Sky-view factor as a relief visualization technique}
\usage{
skyViewFactor(path.software, path.input, options_overwrite = 1,
  options_exaggeration_factor = 1, hillshading = c(0, 315, 35, 0),
  multiple_hillshading = c(0, 16), pca_hillshading = c(0, 3),
  slope_gradient = 0, simple_local_relief = c(0, 20),
  sky_view_factor = c(1, 16, 10, 0, "low"), anisotropic_svf = c(0,
  "low", 315), positive_openness = 0, negative_openness = 0,
  sky_illumination = c(0, "overcast", 250, 100), local_dominance = c(0,
  10, 20), quiet = TRUE)
}
\arguments{
\item{path.software}{path where software is located}

\item{path.input}{path to input file}

\item{quiet}{option to control outputs in console during process. Default: TRUE}

\item{other...}{see \href{http://iaps.zrc-sazu.si/sites/default/files/rvt_1.3_0.pdf}{Relief Visualization Toolbox} \emph{(last call: 13-04-2017)}}
}
\description{
This function enables the possibility to directly address the
stand-alone Relief Visualization Toolbox of ZAKSEK et al. (2011).
}
\note{
\itemize{
  \item output is located in input directory
  \item compatible to Relief Visualization Toolbox version 1.3, \href{http://iaps.zrc-sazu.si/en/rvt#v}{RVT 1.3} \emph{(last call: 13-04-2017)}
  \item ZAKSEK, K., K. OSTIR & Z. KOKALJ (2011): Sky-View Factor as a Relief Visualization Technique. - Remote Sensing 3, 12, 398-415.
}
}
\keyword{factor}
\keyword{sky-view}
