% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_utils.R
\name{st_bbox_geom}
\alias{st_bbox_geom}
\title{Return geometry from bounding box}
\usage{
st_bbox_geom(x, extent = NULL)
}
\arguments{
\item{x}{object of class sf}

\item{extent}{vector containing numeric values, in the following order: xmin, xmax, ymax, ymin}
}
\value{
Geometry of class sfc
}
\description{
This function calculate the geometry based on a bounding box.
}
\keyword{bounding}
\keyword{box}
\keyword{feature,}
\keyword{geometry}
\keyword{of}
\keyword{simple}
