% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_utils.R
\name{st_dissolve}
\alias{st_dissolve}
\title{Dissolve geometry}
\usage{
st_dissolve(x, by = NULL, ...)
}
\arguments{
\item{x}{object of class sf}

\item{by}{field for dissolving. Default: NULL}

\item{...}{Optional: field statistcs}
}
\value{
dissolved geometry of class sf
}
\description{
This function dissovle a geometry based on a field. Optionnally, field statistics can be computed.
}
\keyword{dissolve}
\keyword{feature,}
\keyword{simple}
