% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_utils.R
\name{st_erase}
\alias{st_erase}
\title{Erase one geometry from another}
\usage{
st_erase(x, y)
}
\arguments{
\item{x}{object of class sf}

\item{y}{object of class sf}
}
\value{
Geometry of class sfc
}
\description{
This function erase one geometry from another. The projection must be identical.
}
\keyword{erase}
\keyword{feature,}
\keyword{simple}
