% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_utils.R
\name{st_integration_index}
\alias{st_integration_index}
\title{Calculate integration index}
\usage{
st_integration_index(geom.old, geom.new, geom.boundary = NULL,
  tol = 0.1, precision = 0, ignr.overlap = FALSE, env.rsaga = NULL,
  use.saga = FALSE, return.geom = FALSE, quiet = FALSE)
}
\arguments{
\item{geom.old}{object of class sf representing a land use class of a previous time step}

\item{geom.new}{object of class sf representing a land use class of a following time step}

\item{geom.boundary}{polygon of class sf representing subregions, e.g. administrative boundaries}

\item{tol}{tolerance value for overlapping area m square}

\item{precision}{precision for process. See sf::st_set_precision(). Default: 0}

\item{env.rsaga}{environment of SAGA GIS. If st_erase fails then SAGA GIS erase is used. Default: NULL, but in function call if not set: RSAGA::rsaga.env()}

\item{use.saga}{use SAGA GIS for erase process. Default: FALSE}

\item{return.geom}{If set to TRUE, intermediate geometries are returned as well. Default: FALSE}

\item{quiet}{show output on console. Default: FALSE}
}
\value{
Vector with integration index (completly integrated: 2/3 < R < 1; good integrated: 1/3 < R < 2/3; low integrated: 0 < R < 1/3; not integrated: 0)
}
\description{
This function calculates an index defining the intergration of new land use area into conisting land use area
of the same land use class conisdering two different time steps.
}
\keyword{edge}
\keyword{feature,}
\keyword{index,}
\keyword{interior}
\keyword{ratio}
\keyword{shape}
\keyword{simple}
