% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_utils.R
\name{st_make_grid_lines}
\alias{st_make_grid_lines}
\title{Make multiline fishnet}
\usage{
st_make_grid_lines(x, cellsize)
}
\arguments{
\item{x}{object of class sf}
}
\value{
Geometry of class sfc
}
\description{
This function creates a fishnet, consisting of multilines in x and y direction.
}
\keyword{feature,}
\keyword{fishnet,}
\keyword{multilines}
\keyword{simple}
