% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_utils.R
\name{st_mesh}
\alias{st_mesh}
\title{Calculate the effective mesh size}
\usage{
st_mesh(geom.frag, geom.boundary = NULL, total.area = NULL,
  conv = 10000, do.preProcessing = TRUE, return.geom = FALSE,
  quiet = TRUE)
}
\arguments{
\item{geom.frag}{polygon of class sf representing the fragmentation geometry}

\item{geom.boundary}{polygon of class sf representing subregions, e.g. administrative boundaries}

\item{total.area}{Numeric value representing size for area. Only to use if geom.boundary is not present. Value must match with the conversion constant c (default hectare). Default: NULL}

\item{conv}{constant to convert original square m output. Default: 10000 to convert to hectare. If set to 1, than meter square is the result.}

\item{do.preProcessing}{If TRUE (default), the input of geom.frag is, first, dissolved to single part feature, and second, splitted to multi-parts. By this step it is assured, that polygon connected to each other are summarized}

\item{return.geom}{If set to TRUE, intermediate geometries are returned as well. Default: FALSE}

\item{quiet}{If set to FALSE, actual state is printed to console. Default: TRUE.}
}
\value{
Vector with shape index and interior edge ratio values
}
\description{
This function calculates the effective mesh size.
}
\keyword{effective}
\keyword{feature,}
\keyword{mesh}
\keyword{mesh,}
\keyword{simple}
\keyword{size}
