% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_utils.R
\name{st_perimeter}
\alias{st_perimeter}
\title{Calculate the perimeter of a polygon}
\usage{
st_perimeter(x, ...)
}
\arguments{
\item{x}{object of class sf}
}
\value{
Vector with perimeter values
}
\description{
This function calculates the perimeter of a polygon.
}
\keyword{feature,}
\keyword{perimeter}
\keyword{simple}
