% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_utils.R
\name{st_plot}
\alias{st_plot}
\title{Plot sf geometry}
\usage{
st_plot(x)
}
\arguments{
\item{x}{object of class sf}
}
\value{
plot
}
\description{
This function simply plots the geometry of a sf object.
}
\keyword{feature,}
\keyword{plot}
\keyword{simple}
