% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_utils.R
\name{st_rbind}
\alias{st_rbind}
\title{Row bind simple features}
\usage{
st_rbind(x, geom)
}
\arguments{
\item{x}{list of objects of class sf}

\item{geom}{vector containing the geometries of x. Order must be the same as in list.}
}
\value{
sf object with combined data.frames
}
\description{
This function combines simple feature with different columns using dplr::bind_rows().
}
\keyword{feature,}
\keyword{rbind}
\keyword{simple}
