% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dividePolygon.R
\name{dividePolygon}
\alias{dividePolygon}
\title{Devide Polygon into N Equal Area Parts}
\usage{
dividePolygon(x, angle = 0, n, trans = TRUE, btol = 1e-04,
  pgtol = 1e-09, maxit = 100, cores = 1, quiet = TRUE)
}
\arguments{
\item{x}{object of class sf}

\item{angle}{Can be one single value, or a vector with the same length as the input object x. North is 0 degree, and East 90. Default: 0 degree for North-South. For East-West use 90 degree.}

\item{n}{number of parts}

\item{trans}{angle input is transformed, so that N-S cutting is appropriate. Default: TRUE.}

\item{btol}{tolerance value for lower and upper boundary in optim().Default: 0.0001}

\item{pgtol}{tolerance value to reduce the sum of residuals in optimizing function}

\item{maxit}{maximum iteration to convergence in finding optimal points. Default: 100}

\item{cores}{number of cores for parallel processing. Default: 1 (sequential)}

\item{quiet}{no outputs in console. Default: TRUE}
}
\value{
Geometry similar to sf input including divisions.
}
\description{
This function divides a polygon feature into n-party of approximatly the same area.
The division is performed along the y-axis (N-S). Therefore, polygons should be rotated
appropriate.
}
\note{
\itemize{
  \item Algorithm does not find optimum if polygon split results in more objects than the given number
         n. This can happen when a polygon is very ragged.
  \item Algorithm cut polygon on the N-S-Axis (0-180 degree). Therefore, trans is used to adapt the rotation angle
        so that cutting is appropriate.

}
}
\keyword{appraoch}
\keyword{automatic}
\keyword{event,}
\keyword{landslide,}
\keyword{rainfall}
\keyword{tresholds,}
