% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRainEventData.R
\name{getRainEventData}
\alias{getRainEventData}
\title{Get rain event of a landslide}
\usage{
getRainEventData(x, type = "daily", rainThresh = 0.2, rainOffLength = 2,
  RD = NULL, MAP = NULL, RDN = MAP/RD)
}
\arguments{
\item{x}{vector containing precipitation}

\item{type}{type of precipitation data, i.e. "daily" or "hourly". Default: "daily"}

\item{rainThresh}{rainfall below this threshold is considered as dry. This threshold should be adapted according to type. Default: 0.2}

\item{rainOffLength}{length of dry period to separate rainfall events. This parameter should be adapted according to type. Default: 2 (days), for "hourly" data 48-96 h can be found in literature}

\item{RD}{average number of rainy days in a year, proxy for locate climate conditions. Default: NULL}

\item{MAP}{mean annual precipitation, the long-term yearly average precipitation, see CRU - climate research units for number. Default: NULL}

\item{RDN}{a climatic index that provides better description (or proxy) than the MAP for the occurence of extreme storm events (Guzzetti et al. 2006: 247). Default: MAP/RD}
}
\value{
vector containing rain characeristics (see description).
}
\description{
This function calcuates different precipitation characteristics for a specific time-series:
total precipitation, number of rainfall events, weighted mean intensitiy of rainfall events (normalized by MAP, RD or RDN),
cumulative critical event rainfall (normalized by MAP, RD or RDN), maximum rainfall during critical rainfall event,
duration of critical rainfall event, critical rainfall intensitiy (normalized by MAP, RD or RDN), rainfall at day of failure (start date),
rainfall intensity at day of failure (start date), maximum rainfall at day of failure (start date).
}
\note{
\itemize{
  \item Guzzetti, F., Peruccacci, S., Rossi, M., & Stark, C. P. (2007). Rainfall thresholds for the initiation of landslides in central and southern Europe. Meteorology and atmospheric physics, 98(3-4), 239-267.
  \item Rossi, M., Luciani, S., Valigi, D., Kirschbaum, D., Brunetti, M. T., Peruccacci, S., & Guzzetti, F. (2017). Statistical approaches for the definition of landslide rainfall thresholds and their uncertainty using rain gauge and satellite data. Geomorphology, 285, 16-27.
}
}
\keyword{characeristics}
\keyword{event,}
\keyword{landslide,}
\keyword{rainfall}
\keyword{tresholds,}
