% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_utils.R
\name{st_urban_sprawl}
\alias{st_urban_sprawl}
\title{Calculate urban sprawl}
\usage{
st_urban_sprawl(geom.urban, geom.boundary = NULL, dist = c(100, 100),
  extent = NULL, force.extent = FALSE, do.preProcessing = TRUE,
  return.geom = FALSE, quiet = TRUE)
}
\arguments{
\item{geom.urban}{polygon of class sf representing the fragmentation geometry}

\item{geom.boundary}{polygon of class sf representing subregions, e.g. administrative boundaries}

\item{dist}{vector containing distance between lines in x and y direction. Default: c(100, 100) [m]}

\item{extent}{Numeric value representing extent for area. Format of vector: c(xmin, xmax, ymax, ymin) . Default: NULL}

\item{force.extent}{If TRUE extent is used instead of geom.boundary (if both are present). Default: FALSE}

\item{do.preProcessing}{If TRUE (default), the input of geom.frag is, first, dissolved to single part feature, and second, splitted to multi-parts. By this step it is assured, that polygon connected to each other are summarized}

\item{return.geom}{If set to TRUE, intermediate geometries are returned as well. Default: FALSE}

\item{quiet}{If set to FALSE, actual state is printed to console. Default: TRUE.}
}
\value{
....
}
\description{
This function calculates the urban sprawl of urban area.
}
\keyword{feature,}
\keyword{simple}
\keyword{sprawl}
\keyword{urban}
